Mujoco params, readme, and moved to examples
