+++ sample_factory_examples/mujoco_examples/test_mujoco.py
@@ -2,7 +2,7 @@ import pytest
 
 from sample_factory.algo.utils.misc import ExperimentStatus
 from sample_factory.cfg.arguments import parse_args
-from sample_factory.envs.mujoco.mujoco_utils import mujoco_available
+from sample_factory_examples.mujoco_examples.mujoco_utils import mujoco_available
 from sample_factory.train import run_rl
 from sample_factory.utils.utils import log
 
@@ -44,7 +44,7 @@ class TestMujoco:
         status = run_rl(cfg)
         assert status == ExperimentStatus.SUCCESS
 
-    @pytest.mark.parametrize("env_name", ["mujoco_ant", "mujoco_halfcheetah", "mujoco_humanoid"])
+    @pytest.mark.parametrize("env_name", ["mujoco_ant", "mujoco_halfcheetah", "mujoco_humanoid", "mujoco_hopper", "mujoco_reacher", "mujoco_walker", "mujoco_swimmer"])
     @pytest.mark.parametrize("batched_sampling", [False, True])
     def test_basic_envs(self, env_name, batched_sampling):
         self._run_test_env(env=env_name, num_workers=1, batched_sampling=batched_sampling)
@@ -54,13 +54,3 @@ class TestMujoco:
     def test_single_action_envs(self, env_name, batched_sampling):
         """These envs only have a single action and might cause unique problems with 0-D vs 1-D tensors."""
         self._run_test_env(env=env_name, num_workers=1, train_steps=100)
-
-    @pytest.mark.parametrize("env_name", ["mujoco_hopper", "mujoco_reacher", "mujoco_walker", "mujoco_swimmer"])
-    @pytest.mark.parametrize("batched_sampling", [False, True])
-    def test_rollout_equals_batch_size(self, env_name, batched_sampling):
-        """
-        These envs might create unique problems since their default rollout length is not equal to the batch size.
-        Therefore we have a batch consisting of a single trajectory on the learner and we have to be careful
-        with how we handle it.
-        """
-        self._run_test_env(env=env_name, num_workers=1, batched_sampling=batched_sampling)
